can.operator {operator.tools}	R Documentation

can.operator

Description

tests whether an object can be coerced to a operator, optionally an operator of 'types'.

Usage

can.operator(x, ...)

Arguments

x
object; to test

...
additional arguments

can.operator test whether an object can be coerced to an operator. Methods exist for name, function amd character classes

Value

logical

[Package operator.tools version 1.6.3 Index]