fun2name {operator.tools}	R Documentation

Convert between a function and its name and vice versa.

Description

fun2name compares a function (body) to all defined functions. If an identical match is found to a defined function, that function is returned. NB. This does not search through S4 methods.

Usage

fun2name(f)

name2fun(x)

Arguments

f
function

x
name; more specifically, an object to be converted into a name and eval'd

fun2name compares the function against existing functions using identical. If a match is found, the name of the matching function ( expressed as a character ) is returned.

fun2name will not work for S4 Methods.

Details

name2fun simply converts its argument to a name and than evals it to produce a function definition

Value

fun2name: character (name of function) name2fun: function

[Package operator.tools version 1.6.3 Index]