operator.type {operator.tools}	R Documentation

Return the type for an operator.

Description

Given an operator or its name/symbol, return the type of operator.

Usage

operator.type(op)

Arguments

op
An operator either as a name/symbol or function.

Details

The operator is first checked against all operators that have been registered with the setOperator command. If there is a match, its type is returned. If no matching operator is found, op is matched against unregistered operators that have been defined with the %any%-syntax. If a match is found, UNREGISTERED is returned.

The list of operators are maintained in .Options\$operators and be altered suing the setOperator command.

Value

A character value.

For registered operators, the registered type is returned. For Base R operators, the types come from Syntax.

For operators defined with the %any%-syntax but, not registered using setOperator, "UNREGISTERED" is returned.

NULL is returned otherwise.

Author(s)

Christopher Brown

See Also

operators, setOperator. Syntax

Examples

 ## Not run: 
  operator.type( `+` )
  operator.type( `<=` )
  
  e <- quote( A +B )
  operator.type( e[[1]] )

  operator.type( as.name('+') )
 
## End(Not run)
 

[Package operator.tools version 1.6.3 Index]