removeOperator {operator.tools}	R Documentation

Unregister a an operator.

Description

removeOperator unregisted an operator by removing it from the list of operators found at .Options$operators . All operator attributes are that have been set will be lost.

Usage

removeOperator(x)

Arguments

x
character. The name of the operator

Details

Warns if the operator has not been registered.

Value

None. Used for side-effects.

Author(s)

Christopher Brown

See Also

setOperators for registering Operators.

Examples

  # Unregister ? as an operator.
  ## Not run: 
    removeOperator( '?' )
  
## End(Not run)

[Package operator.tools version 1.6.3 Index]